+++ e2e/runner/setup_sui.go
@@ -123,8 +123,6 @@ func (r *E2ERunner) SuiUpdateGatewayInfoAndTSS() {
 	tssAddr, err := r.ObserverClient.GetTssAddress(r.Ctx, &observertypes.QueryGetTssAddressRequest{})
 	require.NoError(r, err)
 	r.SuiTSSAddress = tssAddr.Sui
-
-	r.Logger.Info("current gateway package ID: %s", r.SuiGateway.PackageID())
 }
 
 // suiSetupDeployerAccount imports a Sui deployer private key using the sui keytool import command
@@ -523,6 +521,7 @@ func (r *E2ERunner) setSuiChainParams(resetNonces bool) error {
 			FastInboundCount:  1,
 			FastOutboundCount: 1,
 		},
+		GasPriceMultiplier: observertypes.DefaultGasPriceMultiplier,
 	}
 	if err := r.ZetaTxServer.UpdateChainParams(chainParams); err != nil {
 		return errors.Wrap(err, "unable to broadcast sui chain params tx")