+++ changelog.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
-## Unreleased
+## Release ReForge
+- zetacored: v37.0.0
+- zetaclientd: v38.0.0
+
+### Breaking Changes
+
+* A new field `gas_price_multiplier` is added to the chain parameters, impacting the query API `/zeta-chain/observer/get_chain_params` and the cosmos message `MsgUpdateChainParams`.
+The `zetacored` binary must be upgraded to trigger chain parameters data migration before running the new `zetaclientd` binary.
 
 ### Features
 
@@ -11,6 +18,7 @@
 * [4492](https://github.com/zeta-chain/node/pull/4492) - add migration script to set authorization for MsgRemoveObserver
 * [4453](https://github.com/zeta-chain/node/pull/4453) - add an option to create dry zetaclients to connect to live networks
 * [4493](https://github.com/zeta-chain/node/pull/4493) - enable ALT in solana inbounds
+* [4485](https://github.com/zeta-chain/node/pull/4485) - move gas price multiplier to zetacore chain parameters
 
 ### Fixes